+++ .editorconfig
@@ -19,4 +19,3 @@ eclint_indent_style = unset
 
 [Dockerfile]
 indent_size = 4
-© 2021 GitHub, Inc.
